API/requirements.txt
fastapi
uvicorn
mysql-connector-python
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp 
opentelemetry-instrumentation-fastapi
#opentelemetry-instrumentation-mysql
opentelemetry-instrumentation-pymysql